backend/requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
python-multipart==0.0.5
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
passlib==1.7.4
python-dotenv==1.0.0
pypi-server==0.1.0